[riscv] Fix custom-descriptors-inlining.js failed
Fix incorrect operand in AllocateStackSlot

Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6928701
Commit-Queue: Yahan Lu (LuYahan) <yahan@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#102337}

+++ src/wasm/baseline/riscv/liftoff-assembler-riscv-inl.h
@@ -2409,12 +2409,12 @@ void LiftoffAssembler::CallBuiltin(Builtin builtin) {
 }
 
 void LiftoffAssembler::AllocateStackSlot(Register addr, uint32_t size) {
-  AddWord(sp, sp, Operand(-size));
+  AddWord(sp, sp, Operand(-static_cast<int64_t>(size)));
   MacroAssembler::Move(addr, sp);
 }
 
 void LiftoffAssembler::DeallocateStackSlot(uint32_t size) {
-  AddWord(sp, sp, Operand(size));
+  AddWord(sp, sp, Operand(static_cast<int64_t>(size)));
 }
 
 void LiftoffAssembler::MaybeOSR() {}
